pyproject.toml
[project]
name = "pants-devcontainer-power"
version = "0.1.0"
description = "MCP tools for Pants build system with devcontainer integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=0.9.0",
    "hypothesis>=6.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "tomli-w>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
pants-devcontainer-power = "src.server:main"

[project.optional-dependencies]
dev = [
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "property: property-based tests using Hypothesis",
    "feature: feature identifier for property tests",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "pytest"
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "A",      # flake8-builtins (prevent shadowing builtins)
    "B",      # flake8-bugbear (common bugs)
    "C4",     # flake8-comprehensions (better comprehensions)
    "C90",    # mccabe (complexity checking)
    "COM818", # flake8-commas (trailing comma enforcement)
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "PLW0406",# pylint: import-self
    "RUF",    # Ruff-specific rules
    "SIM",    # flake8-simplify (code simplification)
    "SLF001", # flake8-self (private member access)
    "UP",     # pyupgrade (modern Python syntax)
    "W",      # pycodestyle warnings
]
ignore = []
